requirements.txt
Bootstrap_Flask==2.2.0
Requests==2.31.0
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv~=1.0.1